Added a "LAUNCHER" suite-format.
This is the first step to enable Benchpress to run non-Bohrium benchmarks.
This change is enough to be able to run them, but Benchpress still relies
on Bohrium to be installed since a ton of stuff is done to/with the configuration
file.

This should also make it a lot easier to fix issue #1.