idl_init: Added a pseudo-result (R=X+Y+Z).
Numpytest only compares two arrays, one is the result from NumPy,
the other a result from Bohrium.

The benchmark returns a tuple (X,Y,Z) which is not compatible,
with the above-mentioned verification "protocol" of numpytest.

Quick-fix is to provide a single array for `numpytest` by producing
a psedo-result consisting of the sum of X, Y, Z.